Add files via upload
added status to issues. Status can be open or closed, when opened users are free to comment, and the issue owner/admins are free to edit the issue. When the issue is closed, users are unable to comment on the issue, and the issue owner/admins cannot edit the issue unless they decide to reopen issue.